read_lines: don't add last line if it's empty
this can happen if there's a newline before EOF: "something\n\0".
If you think of newlines as line terminators that's one line. If
you think of them as line separators, that's two lines. The
convention in Pipes is that's one line, because that leads to
simple algorithms parsing every line with an expected format.