actually maybe we can't avoid the clone
even for replacing, we can't take a &mut because
the inner chain evaluation takes a &mut self. Removing from
lists and inserting again might be a bad idea if we want to
keep the list pointer, and if we access the list inside the loop
(e.g. add current elem plus next elem).